+++ fractal_server/app/schemas/_validators.py
@@ -1,6 +1,4 @@
 import os
-from datetime import datetime
-from datetime import timezone
 from typing import Any
 
 
@@ -100,18 +98,5 @@ def val_unique_list(attribute: str):
             if len(set(must_be_unique)) != len(must_be_unique):
                 raise ValueError(f"`{attribute}` list has repetitions")
         return must_be_unique
-
-    return val
-
-
-def valutc(attribute: str):
-    def val(timestamp: datetime | None) -> datetime | None:
-        """
-        Replacing `tzinfo` with `timezone.utc` is just required by SQLite data.
-        If using Postgres, this function leaves the datetime exactly as it is.
-        """
-        if timestamp is not None:
-            return timestamp.replace(tzinfo=timezone.utc)
-        return None
 
     return val