+++ fractal_server/app/schemas/v1/applyworkflow.py
@@ -6,7 +6,6 @@ from pydantic import validator
 from pydantic.types import StrictStr
 
 from .._validators import valstr
-from .._validators import valutc
 from .dumps import DatasetDumpV1
 from .dumps import ProjectDumpV1
 from .dumps import WorkflowDumpV1
@@ -148,13 +147,6 @@ class ApplyWorkflowReadV1(_ApplyWorkflowBaseV1):
     working_dir_user: str | None
     first_task_index: int | None
     last_task_index: int | None
-
-    _start_timestamp = validator("start_timestamp", allow_reuse=True)(
-        valutc("start_timestamp")
-    )
-    _end_timestamp = validator("end_timestamp", allow_reuse=True)(
-        valutc("end_timestamp")
-    )
 
 
 class ApplyWorkflowUpdateV1(BaseModel):