+++ fractal_server/app/schemas/v1/dataset.py
@@ -7,7 +7,6 @@ from pydantic import validator
 
 from .._validators import val_absolute_path
 from .._validators import valstr
-from .._validators import valutc
 from .dumps import WorkflowTaskDumpV1
 from .project import ProjectReadV1
 from .workflow import WorkflowTaskStatusTypeV1
@@ -149,10 +148,6 @@ class DatasetReadV1(_DatasetBaseV1):
     read_only: bool
     project: ProjectReadV1
     timestamp_created: datetime
-
-    _timestamp_created = validator("timestamp_created", allow_reuse=True)(
-        valutc("timestamp_created")
-    )
 
 
 class DatasetStatusReadV1(BaseModel):