+++ fractal_server/app/schemas/v1/project.py
@@ -4,7 +4,6 @@ from pydantic import BaseModel
 from pydantic import validator
 
 from .._validators import valstr
-from .._validators import valutc
 
 
 __all__ = (
@@ -48,10 +47,6 @@ class ProjectReadV1(_ProjectBaseV1):
 
     id: int
     timestamp_created: datetime
-
-    _timestamp_created = validator("timestamp_created", allow_reuse=True)(
-        valutc("timestamp_created")
-    )
 
 
 class ProjectUpdateV1(_ProjectBaseV1):