+++ fractal_server/app/schemas/v1/state.py
@@ -2,9 +2,6 @@ from datetime import datetime
 from typing import Any
 
 from pydantic import BaseModel
-from pydantic import validator
-
-from fractal_server.app.schemas._validators import valutc
 
 
 class StateRead(BaseModel):
@@ -18,5 +15,3 @@ class StateRead(BaseModel):
     id: int | None
     data: dict[str, Any]
     timestamp: datetime
-
-    _timestamp = validator("timestamp", allow_reuse=True)(valutc("timestamp"))