+++ fractal_server/app/schemas/v2/job.py
@@ -7,7 +7,6 @@ from pydantic import validator
 from pydantic.types import StrictStr
 
 from .._validators import valstr
-from .._validators import valutc
 from .dumps import DatasetDumpV2
 from .dumps import ProjectDumpV2
 from .dumps import WorkflowDumpV2
@@ -99,13 +98,6 @@ class JobReadV2(BaseModel):
     first_task_index: int | None
     last_task_index: int | None
     worker_init: str | None
-
-    _start_timestamp = validator("start_timestamp", allow_reuse=True)(
-        valutc("start_timestamp")
-    )
-    _end_timestamp = validator("end_timestamp", allow_reuse=True)(
-        valutc("end_timestamp")
-    )
 
 
 class JobUpdateV2(BaseModel, extra=Extra.forbid):