+++ fractal_server/app/schemas/v2/project.py
@@ -5,7 +5,6 @@ from pydantic import Extra
 from pydantic import validator
 
 from .._validators import valstr
-from .._validators import valutc
 
 
 class ProjectCreateV2(BaseModel, extra=Extra.forbid):
@@ -20,10 +19,6 @@ class ProjectReadV2(BaseModel):
     id: int
     name: str
     timestamp_created: datetime
-    # Validators
-    _timestamp_created = validator("timestamp_created", allow_reuse=True)(
-        valutc("timestamp_created")
-    )
 
 
 class ProjectUpdateV2(BaseModel, extra=Extra.forbid):