+++ tests/data/example_server_startup/run_server_gunicorn.sh
@@ -1,6 +1,7 @@
 #!/bin/bash
 
-# Create an empty db
+# Create and init db
+createdb fractal-examples-test
 poetry run fractalctl set-db
 
 # Start the server